+++ src/derivatives_trading_portfolio_margin/rest_api/models/modify_um_order_response.rs
@@ -35,9 +35,6 @@ pub struct ModifyUmOrderResponse {
     /// Price.
     #[serde(rename = "price", skip_serializing_if = "Option::is_none")]
     pub price: Option<String>,
-    /// Avg Price.
-    #[serde(rename = "avgPrice", skip_serializing_if = "Option::is_none")]
-    pub avg_price: Option<String>,
     /// Orig Qty.
     #[serde(rename = "origQty", skip_serializing_if = "Option::is_none")]
     pub orig_qty: Option<String>,
@@ -47,9 +44,6 @@ pub struct ModifyUmOrderResponse {
     /// Cum Qty.
     #[serde(rename = "cumQty", skip_serializing_if = "Option::is_none")]
     pub cum_qty: Option<String>,
-    /// Cum Quote.
-    #[serde(rename = "cumQuote", skip_serializing_if = "Option::is_none")]
-    pub cum_quote: Option<String>,
     /// Time In Force.
     #[serde(rename = "timeInForce", skip_serializing_if = "Option::is_none")]
     pub time_in_force: Option<String>,
@@ -95,11 +89,9 @@ impl ModifyUmOrderResponse {
             client_order_id: None,
             modify_id: None,
             price: None,
-            avg_price: None,
             orig_qty: None,
             executed_qty: None,
             cum_qty: None,
-            cum_quote: None,
             time_in_force: None,
             r#type: None,
             reduce_only: None,